Separate services

+++ routes/heros.js
@@ -9,17 +9,26 @@ function authenticateToken(req, res) {
   if (token == null) return res.sendStatus(401);
 
   jwt.verify(token, process.env.SECRET_JWT, (err, user) => {
-    if (err) return res.sendStatus(403);
+    if (err || !user) return res.sendStatus(403);
   });
 }
 router.get('/find-hero/:id', async function (req, res) {
-  authenticateToken(req, res);
-  try {
-    data = await Hero.findOne({ id: req.params.id });
-    res.json(data);
-  } catch (err) {
-    res.send(500, { message: err.message });
-  }
+  if (!authenticateToken(req, res))
+    try {
+      data = await Hero.findOne({ id: req.params.id });
+      res.json(data);
+    } catch (err) {
+      return res.sendStatus(500, { message: err.message });
+    }
+});
+router.get('/get-list-hero', async function (req, res) {
+  if (!authenticateToken(req, res))
+    try {
+      data = await Hero.find();
+      res.json(data);
+    } catch (err) {
+      return res.sendStatus(500, { message: err.message });
+    }
 });
 router.post('/create-hero', async function (req, res) {
   try {
@@ -27,31 +36,51 @@ router.post('/create-hero', async function (req, res) {
     data = await newHero.save();
     res.json(data);
   } catch (err) {
-    res.send(500, { message: err.message });
+    return res.sendStatus(500, { message: err.message });
   }
 });
 router.patch('/update-hero/:id', async function (req, res) {
-  authenticateToken(req, res);
-  try {
-    const result = await Hero.findOneAndUpdate(
-      { id: req.params.id },
-      req.body,
-      {
-        new: true,
-      }
-    );
-    res.json(result);
-  } catch (err) {
-    res.send(500, { message: err.message });
-  }
+  if (!authenticateToken(req, res))
+    try {
+      const result = await Hero.findOneAndUpdate(
+        { id: req.params.id },
+        req.body,
+        {
+          new: true,
+        }
+      );
+      res.json(result);
+    } catch (err) {
+      return res.sendStatus(500, { message: err.message });
+    }
 });
 router.delete('/delete-hero/:id', async function (req, res) {
-  authenticateToken(req, res);
-  try {
-    const data = await Hero.findOneAndDelete({ id: req.params.id });
-    res.send(200);
-  } catch (err) {
-    res.send(500, { message: err.message });
-  }
+  if (!authenticateToken(req, res))
+    try {
+      const data = await Hero.findOneAndDelete({ id: req.params.id });
+      return res.json(data);
+    } catch (err) {
+      return res.sendStatus(500, { message: err.message });
+    }
+});
+
+router.delete('/delete-multi-heros', async function (req, res) {
+  if (!authenticateToken(req, res))
+    try {
+      const data = await Hero.deleteMany({ id: { $in: req.body.id } });
+      res.json(data);
+    } catch (error) {
+      res.sendStatus(500).json({ message: error.message });
+    }
+});
+
+router.post('/add-multi-heros', async function (req, res) {
+  if (!authenticateToken(req, res))
+    try {
+      const data = await Hero.insertMany(req.body.heros);
+      res.json(data);
+    } catch (error) {
+      res.sendStatus(500).json({ message: error.message });
+    }
 });
 module.exports = router;
